+++ postgresql/migration_sqls/payment/000014_payment_settlement_configuration_4.3.sql
@@ -4,10 +4,10 @@
 USE payment_dev;
 
 CREATE TABLE IF NOT EXISTS payment_dev.settlement_configuration (
-  id BIGSERIAL NOT NULL,
+  id SERIAL NOT NULL,
   currency varchar(3) NOT NULL,
-  service_group_id BIGINT NOT NULL,
-  service_id BIGINT NOT NULL,
+  service_group_id INT NOT NULL,
+  service_id INT NOT NULL,
   is_deleted tinyint(1) DEFAULT '0',
   created_timestamp timestamp(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
   last_updated_timestamp timestamp(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
@@ -15,9 +15,9 @@ CREATE TABLE IF NOT EXISTS payment_dev.settlement_configuration (
 );
 
 CREATE TABLE IF NOT EXISTS payment_dev.settlement_resolving_history (
-  id BIGSERIAL NOT NULL,
-  settlement_id BIGINT NOT NULL,
-  order_id BIGINT NOT NULL,
+  id SERIAL NOT NULL,
+  settlement_id INT NOT NULL,
+  order_id INT NOT NULL,
   created_timestamp timestamp(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
   last_updated_timestamp timestamp(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
   PRIMARY KEY (id)
