+++ postgresql/migration_sqls/payment/000017_payment_settlement_resolving_history_4.3_corrected.sql
@@ -2,7 +2,7 @@
 -- SQL in this section is executed when the migration is applied.
 
 USE payment_dev;
-ALTER TABLE payment_dev.settlement_resolving_history ADD COLUMN settlement_configuration_id BIGINT NOT NULL;
+ALTER TABLE payment_dev.settlement_resolving_history ADD COLUMN settlement_configuration_id INT NOT NULL;
 
 
 -- +goose Down
